// libs/examples/src/tree-table/tree-table-example-types.ts
/**
 * @license
 * Copyright 2019 Dynatrace LLC
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { DtIconType } from '@dynatrace/barista-icons';

export class ThreadNode {
  name: string;
  threadlevel?: string;
  totalTimeConsumption?: number;
  blocked: number;
  running: number;
  waiting: number;
  icon?: DtIconType;
  children?: ThreadNode[];
  isShowMore?: boolean;
}

export class ThreadFlatNode {
  name: string;
  threadlevel?: string;
  totalTimeConsumption?: number;
  blocked: number;
  running: number;
  waiting: number;
  icon?: DtIconType;
  level: number;
  expandable: boolean;
  isShowMore?: boolean;
}